+++ src/main/java/proguard/classfile/visitor/ClassCleaner.java
@@ -423,9 +423,10 @@ implements   ClassVisitor,
         {
             visitKotlinDeclarationContainerMetadata(clazz, kotlinClassKindMetadata);
 
-            kotlinClassKindMetadata.typeParametersAccept(clazz, this);
-            kotlinClassKindMetadata.superTypesAccept(    clazz, this);
-            kotlinClassKindMetadata.constructorsAccept(  clazz, this);
+            kotlinClassKindMetadata.typeParametersAccept(                   clazz, this);
+            kotlinClassKindMetadata.superTypesAccept(                       clazz, this);
+            kotlinClassKindMetadata.constructorsAccept(                     clazz, this);
+            kotlinClassKindMetadata.inlineClassUnderlyingPropertyTypeAccept(clazz, this);
         }
 
         @Override