+++ src/test/kotlin/proguard/classfile/util/kotlin/KotlinMetadataInitializerTest.kt
@@ -19,12 +19,15 @@
 package proguard.classfile.util.kotlin
 
 import io.kotest.core.spec.style.FreeSpec
+import io.kotest.matchers.shouldBe
 import io.mockk.spyk
 import io.mockk.verify
 import proguard.classfile.kotlin.visitor.KotlinMetadataVisitor
 import proguard.classfile.kotlin.visitor.ReferencedKotlinMetadataVisitor
 import proguard.classfile.kotlin.visitor.filter.KotlinDeclarationContainerFilter
+import proguard.classfile.visitor.MultiClassVisitor
 import testutils.ClassPoolBuilder
+import testutils.JavaSource
 import testutils.KotlinSource
 
 class KotlinMetadataInitializerTest : FreeSpec({
@@ -76,6 +79,43 @@ class KotlinMetadataInitializerTest : FreeSpec({
                     programClassPool.getClass("Foo"),
                     withArg {
                         it.ownerClassName = "Foo"
+                    }
+                )
+            }
+        }
+    }
+
+    "Given a class with Kotlin 1.4 metadata" - {
+        val (programClassPool, _) = ClassPoolBuilder.fromSource(
+            JavaSource(
+                "TestKotlin1dot4Metadata.java",
+                """
+                    @kotlin.Metadata(
+                        bv = {1, 0, 3},
+                        d1 = {"\u0000\n\n\u0002\u0018\u0002\n\u0002\u0010\u0000\n\u0000\u0018\u00002\u00020\u0001B\u0005¢\u0006\u0002\u0010\u0002"},
+                        d2 = {"LTestKotlin1dot4Metadata;", "", "()V"},
+                        k = 1,
+                        mv = {1, 4, 0}
+                    )
+                    public class TestKotlin1dot4Metadata { }
+                """.trimIndent()
+            )
+        )
+        "Then the metadata should be parsed correctly" {
+            val visitor = spyk<KotlinMetadataVisitor>()
+            programClassPool.classesAccept(
+                MultiClassVisitor(
+                    KotlinMetadataInitializer { _, _ -> },
+                    ReferencedKotlinMetadataVisitor(visitor)
+                )
+            )
+
+            verify {
+                visitor.visitKotlinClassMetadata(
+                    programClassPool.getClass("TestKotlin1dot4Metadata"),
+                    withArg {
+                        it.className shouldBe "TestKotlin1dot4Metadata"
+                        it.mv shouldBe arrayOf(1, 4, 0)
                     }
                 )
             }